requirements.txt
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
beniget==0.4.1
bleach==5.0.1
blinker==1.4
Brotli==1.0.9
certifi==2022.9.14
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
contourpy==1.0.5
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
entrypoints==0.4
executing==1.0.0
fasteners==0.14.1
fastjsonschema==2.16.1
fonttools==4.37.2
fs==2.4.12
future==0.18.2
gast==0.5.2
h11==0.13.0
html5lib==1.1
idna==3.4
importlib-metadata==4.6.4
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyterlab==3.4.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.1
kaggle==1.5.12
keyring==23.5.0
kiwisolver==1.4.4
lab==7.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
nbclassic==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
netifaces==0.11.0
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.3
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.12.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-slugify==6.1.2
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.0
qtconsole==5.3.2
QtPy==2.2.0
reportlab==3.6.8
requests==2.28.1
scipy==1.8.0
SecretStorage==3.3.1
selenium==4.4.3
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.5.0
sympy==1.9
systemd-python==234
terminado==0.15.0
testpath==0.5.0
text-unidecode==1.3
tinycss2==1.1.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trio==0.21.0
trio-websocket==0.9.2
txt2tags==3.7
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.12
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
webdriver-manager==3.8.3
webencodings==0.5.1
websocket-client==1.4.1
widgetsnbextension==4.0.3
wsproto==1.2.0
xdg==5
xlrd==2.0.1
zipp==1.0.0
